ref(onboarding): Split native onboarding docs
