+++ static/app/gettingStartedDocs/native/native/utils.tsx
@@ -1,0 +1,6 @@
+export const getVerifySnippet = () => `
+sentry_capture_event(sentry_value_new_message_event(
+  /*   level */ SENTRY_LEVEL_INFO,
+  /*  logger */ "custom",
+  /* message */ "It works!"
+));`;
