+++ static/app/gettingStartedDocs/native/qt/index.tsx
@@ -1,0 +1,11 @@
+import type {Docs} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {CrashReportWebApiOnboarding} from 'sentry/components/onboarding/gettingStartedDoc/utils/feedbackOnboarding';
+
+import {onboarding} from './onboarding';
+
+const docs: Docs = {
+  onboarding,
+  crashReportOnboarding: CrashReportWebApiOnboarding,
+};
+
+export default docs;
